requirements.txt
asgiref==3.5.2
braintree==4.0
Django==3.2.16
djangorestframework==3.13.0
djangorestframework-simplejwt==5.2.1
drf-writable-nested==0.7.0
python-dotenv==0.21.0
PyYAML==6.0
uritemplate==4.1.1
